personal-ai-assistant/requirements.txt
speechrecognition
pyaudio
pyttsx3
google-generativeai
